Step 1: Initializing the Project
$ kubebuilder init --project-version 2 --domain metamagical.io

Initialize the project with to a new-style project (scaffolding version 2) with a
base domain of `metamagical.io` (my personal domain -- use your own for
reproducing this).

The base domain will form the base of any API groups we have, so if we
created an API group `data`, we'd get a full group name of
`data.metamagical.io`.

This scaffolds out for us a project using Go modules, a Makefile to
build, test, and deploy our project, an initial main file, and kustomize
manifests with different ways to launch our project.