Step 2a: Scaffold GuestBook API
$ kubebuilder create api --group webapp --kind GuestBook --version v1

This will scaffold out empty API types and an empty controller for our
new GuestBook CRD.  The full CRD name will be
guestbooks.webapp.metamagical.io (insert your domain here).

We'll have to fill in the types with the information that we need to
launch our guestbook, but KubeBuilder has provided us with the basics:
a base type and list type, as well as a spec (for desired state) and a
status (for observed state).

We'll also have to fill in the controller, which is now run when our
manager is run.